# docs/swagger.yaml
basePath: /v2/
definitions:
...
  model.Accessible:
    properties:
      canDelete:
        type: boolean
      canEdit:
        type: boolean
      canView:
        type: boolean
      createdAt:
        type: string
      entityId:
        type: string
      id:
        type: string
      isOwner:
        type: boolean
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.Comment:
    properties:
      entityId:
        type: string
      id:
        type: string
      text:
        type: string
      userId:
        type: string
    type: object
...
  model.File:
    properties:
      createdAt:
        type: string
      deploymentType:
        description: server or client if applicable
        type: string
      entityId:
        type: string
      hash:
        type: string
      height:
        type: integer
      id:
        type: string
      mime:
        type: string
      originalPath:
        description: original relative path to maintain directory structure (e.g.
          for releases)
        type: string
      platform:
        description: platform if applicable
        type: string
      size:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
      uploadedBy:
        description: user that uploaded the file
        type: string
      url:
        type: string
      variation:
        description: variant of the file if applicable (e.g. PDF pages)
        type: integer
      version:
        description: version of the file if versioned
        type: integer
      width:
        type: integer
    type: object
  model.Likable:
    properties:
      createdAt:
        type: string
      entityId:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      userId:
        description: User who liked or disliked the Entity
        type: string
      value:
        description: 1 for likes, -1 for dislikes
        type: integer
    type: object
  model.Link:
    properties:
      id:
        type: string
      name:
        type: string
      url:
        type: string
    type: object
  model.Package:
    properties:
      accessibles:
        items:
          $ref: '#/definitions/model.Accessible'
        type: array
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      createdAt:
        type: string
      description:
        type: string
      downloads:
        type: integer
      entityType:
        type: string
      files:
        items:
          $ref: '#/definitions/model.File'
        type: array
      id:
        type: string
      likables:
        items:
          $ref: '#/definitions/model.Likable'
        type: array
      liked:
        type: integer
      links:
        items:
          $ref: '#/definitions/model.Link'
        type: array
      map:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/model.User'
      price:
        type: number
      properties:
        items:
          $ref: '#/definitions/model.Property'
        type: array
      public:
        type: boolean
      release:
        type: string
      releasedAt:
        type: string
      summary:
        type: string
      title:
        type: string
      totalDislikes:
        type: integer
      totalLikes:
        type: integer
      updatedAt:
        type: string
      version:
        type: string
      views:
        type: integer
    type: object
  model.Persona:
    properties:
      accessibles:
        items:
          $ref: '#/definitions/model.Accessible'
        type: array
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      configuration: {}
      createdAt:
        type: string
      entityType:
        type: string
      files:
        items:
          $ref: '#/definitions/model.File'
        type: array
      id:
        type: string
      likables:
        items:
          $ref: '#/definitions/model.Likable'
        type: array
      links:
        items:
          $ref: '#/definitions/model.Link'
        type: array
      name:
        type: string
      owner:
        $ref: '#/definitions/model.User'
      properties:
        items:
          $ref: '#/definitions/model.Property'
        type: array
      public:
        type: boolean
      type:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      views:
        type: integer
    type: object
...
  model.Presence:
    properties:
      serverId:
        type: string
      spaceId:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  model.Property:
    properties:
      entityId:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      value:
        type: string
    type: object
  model.User:
    properties:
      accessibles:
        items:
          $ref: '#/definitions/model.Accessible'
        type: array
      activatedAt:
        type: string
      address:
        type: string
      allowEmails:
        type: boolean
      apiKey:
        type: string
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      createdAt:
        type: string
      defaultPersona:
        $ref: '#/definitions/model.Persona'
      description:
        type: string
      email:
        type: string
      entityType:
        type: string
      ethAddress:
        type: string
      experience:
        type: integer
      files:
        items:
          $ref: '#/definitions/model.File'
        type: array
      id:
        type: string
      isActive:
        type: boolean
      isAdmin:
        type: boolean
      isBanned:
        type: boolean
      isInternal:
        type: boolean
      isMuted:
        type: boolean
      lastSeenAt:
        type: string
      level:
        type: integer
      likables:
        items:
          $ref: '#/definitions/model.Likable'
        type: array
      links:
        items:
          $ref: '#/definitions/model.Link'
        type: array
      name:
        type: string
      owner:
        $ref: '#/definitions/model.User'
      presence:
        $ref: '#/definitions/model.Presence'
      properties:
        items:
          $ref: '#/definitions/model.Property'
        type: array
      public:
        type: boolean
      rank:
        type: string
      updatedAt:
        type: string
      views:
        type: integer
    type: object
  model.World:
    properties:
      accessibles:
        items:
          $ref: '#/definitions/model.Accessible'
        type: array
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      createdAt:
        type: string
      description:
        type: string
      entityType:
        type: string
      files:
        items:
          $ref: '#/definitions/model.File'
        type: array
      gameMode:
        type: string
      id:
        type: string
      likables:
        items:
          $ref: '#/definitions/model.Likable'
        type: array
      liked:
        type: integer
      links:
        items:
          $ref: '#/definitions/model.Link'
        type: array
      map:
        type: string
      metaverse:
        $ref: '#/definitions/model.Package'
      name:
        type: string
      owner:
        $ref: '#/definitions/model.User'
      properties:
        items:
          $ref: '#/definitions/model.Property'
        type: array
      public:
        type: boolean
      totalDislikes:
        type: integer
      totalLikes:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
      views:
        type: integer
    type: object
info:
  contact:
    email: support@le7el.com
    name: Support
  description: LE7EL API swagger
  termsOfService: http://swagger.io/terms/
  title: LE7EL API
  version: 1.0.0
paths:
...
  /worlds/:id:
    get:
      consumes:
      - application/json
      description: Fetch a single world
      parameters:
      - description: ID
        in: query
        name: id
        type: string
      - description: Specify to attach the pak file
        in: query
        name: platform
        type: string
      - description: Specify to attach the pak file
        in: query
        name: deployment
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.World'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - Bearer: []
      summary: Get World
      tags:
      - worlds
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"